+++ packages/runtime/src/index.ts
@@ -1,4 +1,6 @@
 export * from "./RuntimeEvent";
 export * from "./RuntimeHost";
 export * from "./RuntimeModule";
+export * from "./RuntimeModuleSnapshot";
+export * from "./RuntimeModuleStatus";
 export * from "./RuntimeServiceKeys";